Refactored the client so it is split into more components and operates a little bit more cleanly.
