+++ src/components/client/simulatorData.js
@@ -1,0 +1,72 @@
+import React, { Component } from "react";
+import { Query } from "react-apollo";
+import gql from "graphql-tag";
+import SubscriptionHelper from "../../helpers/subscriptionHelper";
+import Client from "./client";
+import playSound from "../generic/SoundPlayer";
+
+const queryData = `
+id
+name
+alertlevel
+layout
+bridgeOfficerMessaging
+assets {
+  mesh
+  texture
+  side
+  top
+  logo
+}
+`;
+
+const QUERY = gql`
+  query Simulator($simulatorId: String!) {
+    simulators(id: $simulatorId) {
+${queryData}
+    }
+  }
+`;
+const SUBSCRIPTION = gql`
+  subscription SimulatorUpdate($simulatorId: ID!) {
+    simulatorsUpdate(simulatorId: $simulatorId) {
+${queryData}
+    }
+  }
+`;
+
+class SimulatorData extends Component {
+  state = {};
+  componentDidMount() {
+    this.props.playSound({ url: "/sciences.ogg" });
+  }
+  render() {
+    return (
+      <Query query={QUERY} variables={{ simulatorId: this.props.simulator.id }}>
+        {({ loading, data, subscribeToMore }) => {
+          const { simulators } = data;
+          if (loading || !simulators) return null;
+          if (!simulators[0]) return <div>No Simulator</div>;
+          return (
+            <SubscriptionHelper
+              subscribe={() =>
+                subscribeToMore({
+                  document: SUBSCRIPTION,
+                  variables: { simulatorId: this.props.simulator.id },
+                  updateQuery: (previousResult, { subscriptionData }) => {
+                    return Object.assign({}, previousResult, {
+                      simulator: subscriptionData.data.simulatorsUpdate
+                    });
+                  }
+                })
+              }
+            >
+              <Client {...this.props} simulator={simulators[0]} />
+            </SubscriptionHelper>
+          );
+        }}
+      </Query>
+    );
+  }
+}
+export default playSound(SimulatorData);
