+++ src/containers/config.scss
@@ -1,7 +1,7 @@
 .config-container {
   height: 100%;
   background-color: #111;
-  background: url(../components/credit-bg.jpg);
+  background: url(../components/client/credit-bg.jpg);
   color: white;
   .card {
     border: solid 1px rgba(255, 255, 255, 0.1);
