+++ src/translations/src/components/client/Card.json
@@ -1,0 +1,6 @@
+[
+  {
+    "id": "sound-player",
+    "defaultMessage": "Sound Player"
+  }
+]
